refactor: 更换`skip`方式.
